Refactor submissions
