--- src/pages/Submissions/telaProdutor.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import Event from '../Events/Event'
-import SubmissionsApproved from './submissionsApproved'
-import SubmissionInAnalysis from './submission'
-
-const TelaProdutor = () => {
-
-  return (
-    <>
-      <Event />
-      <SubmissionsApproved />
-      <SubmissionInAnalysis />
-    </>
-  )
-}
-
-export default TelaProdutor
